+++ test/PDB/AlphaFoldDB.jl
@@ -2,13 +2,13 @@
     @testset "query_alphafolddb tests" begin
         @testset "Valid UniProt ID" begin
             structure_info = query_alphafolddb("A0A0C5B5G6")
-            @test structure_info[1]["uniprotAccession"] == "A0A0C5B5G6"
-            @test structure_info[1]["uniprotId"] == "MOTSC_HUMAN"
-            @test match(r"^https.+\.pdb$", structure_info[1]["pdbUrl"]) !== nothing
+            @test structure_info["uniprotAccession"] == "A0A0C5B5G6"
+            @test structure_info["uniprotId"] == "MOTSC_HUMAN"
+            @test match(r"^https.+\.pdb$", structure_info["pdbUrl"]) !== nothing
         end
 
-        # @testset "Invalid UniProt ID" begin
-        #     @test_throws HTTP.Exceptions.StatusError query_alphafolddb("INVALID_ID")
+        # @testset "Invalid UniProt Accession" begin
+        #     @test_throws HTTP.Exceptions.StatusError query_alphafolddb("INVALID_ACCESSION")
         # end
     end
 
@@ -18,6 +18,18 @@
             if isfile(outfile)
                 try
                     res = read_file(outfile, PDBFile)
+                    @test length(res) == 16
+                finally
+                    rm(outfile)
+                end
+            end
+        end
+
+        @testset "Download mmCIF format" begin
+            outfile = download_alphafold_structure("A0A0C5B5G6", format = MMCIFFile)
+            if isfile(outfile)
+                try
+                    res = read_file(outfile, MMCIFFile)
                     @test length(res) == 16
                 finally
                     rm(outfile)